pyproject.toml
[project]
name = "reverse-loop"
version = "0.1.0"
description = "Autonomous Reverse Logistics Agent Swarm"
requires-python = ">=3.10"

dependencies = [
    "streamlit>=1.30",
    "google-genai",
    "google-adk>=1.1.0",
    "python-dotenv>=1.0",
    "pydantic>=2.0",
    "pillow>=10.0",
    "watchdog>=4.0"
]

[tool.setuptools.packages.find]
where = ["."]
